requirements.txt
# Core Flask dependencies
Flask==3.0.3
gunicorn==22.0.0
python-dotenv==1.0.1

# Database and ORM
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication and security
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Limiter==3.6.0

# Google Cloud services
google-cloud-storage==2.16.0
google-cloud-tasks==2.16.0
google-cloud-secret-manager==2.19.0
google-cloud-documentai==2.27.0
google-auth==2.23.4
google-api-core==2.15.0

# File processing and utilities
filetype==1.2.0
requests==2.31.0
reportlab==4.0.4

# Monitoring and observability
sentry-sdk[flask]==1.40.0